RE #4313

Merged-on: https://opencell.assembla.com/code/meveo/git/merge_requests/8121259

+++ opencell-admin/ejbs/src/main/java/org/meveo/admin/job/MediationJobBean.java
@@ -183,7 +183,7 @@ public class MediationJobBean {
             }
 
             // Delete reject file if it is empty
-            if (result.getErrors().isEmpty() && rejectFile != null) {
+            if ((result.getErrors().isEmpty() && result.getNbItemsProcessedWithError() == 0) && rejectFile != null) {
                 try {
                     rejectFile.delete();
                 } catch (Exception e) {
